+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add support for batch sentiment analysis
+
 ## [1.9.0] - 2023-06-07
 
 ### Fixed